requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
secure-smtplib==0.1.1
setuptools==73.0.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.3
wheel==0.44.0
